requirements.txt
anyio==3.5.0
argcomplete==1.12.3
base58==2.1.1
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.4.3
blivet-gui==2.3.0
Brlapi==0.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
chrome-gnome-shell==0.0.0
click==8.0.1
cryptography==3.4.7
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
distro==1.6.0
ECPy==1.2.5
fedora-third-party==0.8
fros==1.1
gpg==1.15.1
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
humanize==0.5.1
idna==3.2
langtable==0.0.57
libcomps==0.1.18
lxml==4.6.5
Mako==1.1.4
MarkupSafe==2.0.0
nftables==0.1
numpy==1.22.3
olefile==0.46
packaging==21.3
pandas==1.4.2
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==8.3.2
ply==3.11
productmd==1.33
ptyprocess==0.6.0
pwquality==1.4.4
pycairo==1.20.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.44.1
pyenchant==3.2.2
PyGObject==3.42.0
pykickstart==3.34
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.1
python-distutils-extra==2.39
python-meh==0.50
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
regex==2022.3.2
requests==2.27.0
requests-file==1.5.1
requests-ftp==0.3.1
rfc3986==1.5.0
rpm==4.17.0
selinux==3.3
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.8.2
six==1.16.0
sniffio==1.2.0
sos==4.2
soupsieve==2.3.1
systemd-python==234
Tempita==0.5.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websockets==10.2
xrpl-py==1.4.0
